+++ SERVER.CLASS.BUILDER/Server_Header_File_Builder/Server_Header_File_Builder.cpp
@@ -67,11 +67,11 @@ void Server_Header_File_Builder::Determine_Inclusion_Word(){
 
      char * Server_Class_Name = this->Reader_Pointer->Get_Server_Class_Name();
 
-     int WordSize = strlen(Server_Class_Name);
+     size_t WordSize = strlen(Server_Class_Name);
 
      this->InclusionWord = new char [10*WordSize];
 
-     for(int i=0;i<WordSize;i++){
+     for(size_t i=0;i<WordSize;i++){
 
         this->InclusionWord[i] = this->ConvertToUpper(Server_Class_Name[i]);
      }
@@ -261,15 +261,15 @@ void Server_Header_File_Builder::Move_Header_File(){
 
      char Directory_Character [] = {'\\','\0'};
 
-     int Header_File_Name_Size = strlen(Header_File_Name);
+     size_t Header_File_Name_Size = strlen(Header_File_Name);
 
-     int New_Include_Directory_Name_Size = strlen(this->New_Include_Directory);
+     size_t New_Include_Directory_Name_Size = strlen(this->New_Include_Directory);
 
-     int Construction_Point_Name_Size = strlen(this->Construction_Point);
+     size_t Construction_Point_Name_Size = strlen(this->Construction_Point);
 
-     int Current_Path_Name_Size = Construction_Point_Name_Size + Header_File_Name_Size;
+     size_t Current_Path_Name_Size = Construction_Point_Name_Size + Header_File_Name_Size;
 
-     int New_Path_Name_Size = New_Include_Directory_Name_Size + Header_File_Name_Size;
+     size_t New_Path_Name_Size = New_Include_Directory_Name_Size + Header_File_Name_Size;
 
      char * Current_Path_Name = new char [10*Current_Path_Name_Size];
 
@@ -304,9 +304,9 @@ void Server_Header_File_Builder::Move_Header_File(){
 
 void Server_Header_File_Builder::Place_Information(char ** Pointer, char * Information, int * counter){
 
-     int Information_Size = strlen(Information);
+     size_t Information_Size = strlen(Information);
 
-     for(int i=0;i<Information_Size;i++){
+     for(size_t i=0;i<Information_Size;i++){
 
          (*Pointer)[(*counter)] = Information[i];
