+++ SERVER.CLASS.BUILDER/Thread_Locker_Builder/Thread_Locker_Builder.cpp
@@ -177,21 +177,21 @@ void Thread_Locker_Builder::Determine_Compiler_Command(){
 
      char Error_Message_File_Name [] = {'\\','C','o','m','p','i','l','e','r','_','O','u','t','p','u','t','\0'};
 
-     int Source_File_Name_Size = strlen(Source_File_Name);
+     size_t Source_File_Name_Size = strlen(Source_File_Name);
 
-     int Header_File_Name_Size = strlen(Header_File_Name);
+     size_t Header_File_Name_Size = strlen(Header_File_Name);
 
-     int Process_Command_Name_Size = strlen(Process_Command);
+     size_t Process_Command_Name_Size = strlen(Process_Command);
 
-     int Thread_Library_Name_Size = strlen(Thread_Library_Name);
+     size_t Thread_Library_Name_Size = strlen(Thread_Library_Name);
 
-     int Include_Directory_Determiner_Size = strlen(Include_Directory_Determiner);
+     size_t Include_Directory_Determiner_Size = strlen(Include_Directory_Determiner);
 
-     int Current_Directory_Name_Size = strlen(this->Directory_Manager.GetCurrentlyWorkingDirectory());
+     size_t Current_Directory_Name_Size = strlen(this->Directory_Manager.GetCurrentlyWorkingDirectory());
 
-     int Include_Word_Name_Size = strlen(Include_Word);
+     size_t Include_Word_Name_Size = strlen(Include_Word);
 
-     int Compiler_Command_Size = Source_File_Name_Size + Header_File_Name_Size
+     size_t Compiler_Command_Size = Source_File_Name_Size + Header_File_Name_Size
 
                                 + Process_Command_Name_Size + Thread_Library_Name_Size
 
@@ -255,11 +255,11 @@ void Thread_Locker_Builder::Remove_Header_Extra_File(){
 
      char * Current_Directory = this->Directory_Manager.GetCurrentlyWorkingDirectory();
 
-     int Current_Directory_Size = strlen(Current_Directory);
+     size_t Current_Directory_Size = strlen(Current_Directory);
 
-     int Header_File_Name_Size = strlen(Header_Extra_File);
+     size_t Header_File_Name_Size = strlen(Header_Extra_File);
 
-     int File_Path_Size = Current_Directory_Size  + Header_File_Name_Size;
+     size_t File_Path_Size = Current_Directory_Size  + Header_File_Name_Size;
 
      char * File_Path = new char [10*File_Path_Size];
 
@@ -309,9 +309,9 @@ void Thread_Locker_Builder::Run_System_Commands(){
 
 void Thread_Locker_Builder::Place_Information(char ** Pointer, char * Information, int * Counter){
 
-     int Information_Size = strlen(Information);
+     size_t Information_Size = strlen(Information);
 
-     for(int i=0;i<Information_Size;i++){
+     for(size_t i=0;i<Information_Size;i++){
 
         (*Pointer)[(*Counter)] = Information[i];
