+++ SMART.POINTER.BUILDER/Dynamic_Memory_Manager_Header_Builder/Dynamic_Memory_Manager_Header_Builder.cpp
@@ -141,11 +141,11 @@ char Dynamic_Memory_Manager_Header_Builder::ConvertToUpper(char character){
 
 void Dynamic_Memory_Manager_Header_Builder::Determine_Inclusion_Word(){
 
-     int WordSize = strlen(this->Initializer.Get_NewClassName());
+     size_t WordSize = strlen(this->Initializer.Get_NewClassName());
 
      this->InclusionWord = new char [10*WordSize];
 
-     for(int i=0;i<WordSize;i++){
+     for(size_t i=0;i<WordSize;i++){
 
         this->InclusionWord[i] = this->ConvertToUpper(this->Initializer.Get_NewClassName()[i]);
      }
