+++ SMART.POINTER.BUILDER/Multi_Thread_Pointer_Header_Builder/Multi_Thread_Pointer_Header_Builder.cpp
@@ -179,11 +179,11 @@ char Multi_Thread_Pointer_Header_Builder::ConvertToUpper(char character){
 
 void Multi_Thread_Pointer_Header_Builder::Determine_Inclusion_Word(){
 
-     int WordSize = strlen(this->Initializer.Get_NewClassName());
+     size_t WordSize = strlen(this->Initializer.Get_NewClassName());
 
      this->InclusionWord = new char [10*WordSize];
 
-     for(int i=0;i<WordSize;i++){
+     for(size_t i=0;i<WordSize;i++){
 
         this->InclusionWord[i] = this->ConvertToUpper(this->Initializer.Get_NewClassName()[i]);
      }
